+++ flatlaf-core/src/main/java/com/formdev/flatlaf/IntelliJTheme.java
@@ -34,6 +34,7 @@ import javax.swing.UIDefaults;
 import javax.swing.plaf.ColorUIResource;
 import com.formdev.flatlaf.json.Json;
 import com.formdev.flatlaf.json.ParseException;
+import com.formdev.flatlaf.util.LoggingFacade;
 import com.formdev.flatlaf.util.StringUtils;
 
 /**
