+++ flatlaf-core/src/main/java/com/formdev/flatlaf/util/JavaCompatibility.java
@@ -22,7 +22,6 @@ import java.awt.Graphics2D;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import javax.swing.JComponent;
-import com.formdev.flatlaf.LoggingFacade;
 
 /**
  * Provides Java version compatibility methods.
@@ -56,7 +55,7 @@ public class JavaCompatibility
 						? new Class[] { JComponent.class, Graphics2D.class, String.class, int.class, float.class, float.class }
 						: new Class[] { JComponent.class, Graphics.class, String.class, int.class, int.class, int.class } );
 				} catch( Exception ex ) {
-					LoggingFacade.logSevere( ex );
+					LoggingFacade.logSevere( null, ex );
 					throw new RuntimeException( ex );
 				}
 			}
@@ -68,7 +67,7 @@ public class JavaCompatibility
 			else
 				drawStringUnderlineCharAtMethod.invoke( null, c, g, text, underlinedIndex, x, y );
 		} catch( IllegalAccessException | IllegalArgumentException | InvocationTargetException ex ) {
-			LoggingFacade.logSevere( ex );
+			LoggingFacade.logSevere( null, ex );
 			throw new RuntimeException( ex );
 		}
 	}
@@ -92,7 +91,7 @@ public class JavaCompatibility
 							: "clipStringIfNecessary",
 						new Class[] { JComponent.class, FontMetrics.class, String.class, int.class } );
 				} catch( Exception ex ) {
-					LoggingFacade.logSevere( ex );
+					LoggingFacade.logSevere( null, ex );
 					throw new RuntimeException( ex );
 				}
 			}
@@ -101,7 +100,7 @@ public class JavaCompatibility
 		try {
 			return (String) getClippedStringMethod.invoke( null, c, fm, string, availTextWidth );
 		} catch( IllegalAccessException | IllegalArgumentException | InvocationTargetException ex ) {
-			LoggingFacade.logSevere( ex );
+			LoggingFacade.logSevere( null, ex );
 			throw new RuntimeException( ex );
 		}
 	}
