+++ flatlaf-intellij-themes/src/main/java/com/formdev/flatlaf/intellijthemes/Utils.java
@@ -18,7 +18,7 @@ package com.formdev.flatlaf.intellijthemes;
 
 import java.io.IOException;
 import com.formdev.flatlaf.IntelliJTheme;
-import com.formdev.flatlaf.LoggingFacade;
+import com.formdev.flatlaf.util.LoggingFacade;
 
 /**
  * @author Karl Tauber
